requirements.txt
Django==2.0.3
django-oauth-toolkit==1.0.0
djangorestframework==3.7.7
django-cors-middleware==1.3.1
django-extensions==2.0.6
